Consolidate vector embedding tests into SearchIndexAppTest
Merge 3 separate full-app vector embedding test classes
(SearchIndexVectorEmbeddingTest, VectorEmbeddingReindexAppTest,
VectorEmbeddingReembedOperationsTest) into SearchIndexAppTest to avoid
starting infrastructure 3 times. Keep VectorEmbeddingIntegrationIT in
openmetadata-integration-tests since it's self-contained with its own
testcontainers.
